pyproject.toml
[project]
name = "hedrax"
version = "0.1.0"
description = "Polyhedral index helper for static-shape programming in JAX."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "islpy-barvinok>=2025.2.5",
    "jax>=0.7.1",
    "numpy>=2.3.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
